feat: enable all datasets command to operate on dirty repository
