+++ renku/cli/rerun.py
@@ -130,7 +130,10 @@ def edit_inputs(client, workflow):
     nargs=-1,
     required=True,
 )
-@pass_local_client(clean=True, commit=True)
+@pass_local_client(
+    clean=True,
+    commit=True,
+)
 def rerun(client, revision, roots, siblings, inputs, paths):
     """Recreate files generated by a sequence of ``run`` commands."""
     graph = Graph(client)
