+++ renku/models/cwl/__init__.py
@@ -16,7 +16,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Define models for Common Workflow Language."""
-
 from ._ascwl import CWLClass, ascwl
 from .command_line_tool import CommandLineTool
 from .workflow import Workflow
