+++ tests/cli/test_init.py
@@ -15,7 +15,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""CLI tests for init command."""
+"""Test ``init`` command."""
 
 import os
 import shutil
